LU-19472 lod: alloc big enough bitmap for OST avoidance
We use a bitmap to collect used OST indices of other mirrors so that
they are avoided in allocating OST objects in a new mirror. While
it use the number of OSTs for the bitmap size, and when a filesystem
uses sparse OST indices then the OST index would out of the boundary
of the bitmap.

This patch fixes it by using the size of the lu_tgts array as the
OST avoidance bitmap, and this value guarantees to cover all used
OST index (referring to
lov_add_target()->ltd_add_tgt()->lu_tgt_desc_resize())

Test-Parameters: testlist=sanity env=ONLY=103b,ONLY_REPEAT=30
Test-Parameters: testlist=conf-sanity env=ONLY=82c,ONLY_REPEAT=30
Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: Ifdc6b3ecbb278a78df8ca6295921d895e39799d6
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62172
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Feng Lei  <flei@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>