LU-18278 gss: link user keyring to thread keyring
Regular user keys used for GSS contexts are linked to the user
keyring, which is itself linked to the session keyring. This makes
this implemenation fragile to issues with the session keyring, such as
when it is revoked, unlinked, or unavailable. We depend on the kernel
implementation of request_key for that matter, so we need to implement
workarounds.
We choose to resort to the 'thread keyring' trick used elsewhere in
the kernel. This means we make the thread keyring point to the user
keyring when we are dealing with regular user keys. Because of the
implementation of request_key() and search_cred_keyrings_rcu() in the
kernel, we know this is going to be the first keyring being searched.

sanity-krb5 test_13 is added to check this is working as expected.

Test-Parameters: trivial
Test-Parameters: testgroup=review-dne-selinux-ssk-part-1
Test-Parameters: testgroup=review-dne-selinux-ssk-part-2
Test-Parameters: kerberos=true testlist=sanity-krb5
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I43d61dec380fa844d6dbe37b7a96b29d47be3ef2
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62399
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Aurelien Degremont <adegremont@nvidia.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>