LU-19503 doc: Clarify quota limit enforcement behavior
Add documentation noting that Lustre quota limits are
not strictly enforced due to client-side caching, and
should preferably be configured in the GB range.

Reorganize the DESCRIPTION section to have a shorter
section for each command variant, and move the arguments
to a new OPTIONS section instead of repeating them.

Reformat to follow current man page style guidelines.

Test-Parameters: trivial
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Signed-off-by: Andreas Dilger <adilger@thelustrecollective.com>
Change-Id: Ib08e2d74d9c5f6e835167632ffa4d60bd37412d2
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62412
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>