LU-19563 lnet: Increase the default LNET_INTERFACES_MAX_DEFAULT
Increase the default LNET_INTERFACES_MAX_DEFAULT which is used by
lnet_interfaces_max value from 200 to 4096. we are running into
conditions where 200 is not enough especially when there are lots
of vlans. So, lets increase the default lnet_interfaces_max to
another arbitrary high number. The only place it uses are related
to ping and metadata discovery and its only like 160kb or so for 4K
interfaces. The memory pressure is negligible so it should be fine
if we make it future proof with value like 4096. It is mostly used
in lnet_extra_bytes_for_md()

Test-Parameters: trivial
Signed-off-by: Manish Regmi <mregmi@ddn.com>
Change-Id: I08c60fb7ee48872939962681b03db28b99d2604f
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62602
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Chris Horn <chorn@ddn.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Cyril Bordage <cbordage@whamcloud.com>