LU-19827 lnet: Fix MR forwarding
The change https://review.whamcloud.com/57492 introduced a regression
in the multi-rail forwarding code. MR forwarding allows a router to
select a different target interface for a message when the one chosen
by the sender is down or unhealthy.

Prior to the above change, lnet_is_peer_ni_alive() would check both
the down and unhealthy criteria, but now it only checks the down
criteria. We need to add a check for the peer NI's health value.

There was also a second bug here where we were inadvertently checking
the MR capability of the destination peer rather than the source peer.
This is also fixed.

Add sanity-lnet tests to validate MR forwarding feature

Test-Parameters: trivial
Fixes: 70c5f58403 ("LU-18444 lnet: Use only NI status for route aliveness")
Fixes: b0e8ab1a5f ("LU-13606 lnet: Allow router to forward to healthier NID")
HPE-bug-id: LUS-13084
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Idaa40618d1df4c6793266c7cf9ae0d63a5a8bff7
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63582
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>