LU-19625 o2iblnd: handle rdma CONNECT_RESPONSE
RDMA_CM_EVENT_CONNECT_RESPONSE is added to the list of events we
handle to avoid a LBUG. As detailed in the documentation, this event
may be generated on the active side of the connection to notify the
user that the connection request has been successful. The event is
only generated on rdma_cm_ids which do not have a QP associated with
them.
We handle it simply by showing information in the console and ignoring
it.

Signed-off-by: Cyril Bordage <cbordage@whamcloud.com>
Test-Parameters: trivial
Change-Id: Ide168259028849f3ff13d215e0639a735d99fd00
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62679
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>