LU-19136 quota: sync quota setting periodically
The quota setting between QMT and QSD could be out-of-sync
in some case, this patch adds a mechanism to detect and
sync the quota setting, QSD will get a version of global
index with each quota acquire request, and if the last
version update hasn't been updated for 5 minutes, it will
trigger the quota reintergration to get the latest version.

This patch also fixes an issue in test_12a in sanity-quota,
whcih the size to write should take the OST counts into
account because other OSTs could hold the minmum quota grant
(1M) and cause the write to fail.

Signed-off-by: Hongchao zhang <hongchao@whamcloud.com>
Change-Id: Ic9af205ee34056055e51721bbeda04a64bf62ce6
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59948
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>