LU-19079 nodemap: allow switch to gss_id nodemap
If a client was initially part of a nodemap and gss authentication is
in place, the client should be kept as part of this nodemap when the
nodemap is switched to gssonly_identification.
To support this, add a new 'ps_nm_name' field in struct ptlrpc_sec to
store the nodemap name the import is associated with.

sanity-sec test_106 has been added to exercise this.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I6ff91e2f31d8207674e01d444b72179d480c70f8
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63248
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>