LU-19531 mdd: limit mdd_is_parent() looping
Limit the number of iterations that mdd_is_parent() can loop when
trying to find if two FIDs are ancestors of each other.  In some
cases, it is possible for a thread to loop infinitely trying to
walk the tree to the parent FID.

Add debugging to try and identify where the problem is being hit,
whether it is in the cache only, or persistent on disk.

Signed-off-by: Andreas Dilger <adilger@thelustrecollective.com>
Change-Id: Id0338246aa846e14919ddfbea0d058b4a4c98cf6
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62070
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Etienne AUJAMES <eaujames@ddn.com>