LU-12187 llapi: EC layout creation support
Client changes to allow validation of ECed layouts in llapi.

This includes setting up a bi-directional linkage between data and
parity components when a parity component is created. This 1-1
relationship is enforced such that a data component can only be
protected by one parity component.

The linkage between components uses arbitrary IDs such that two
components (data and parity) are linked together by their "link id".
In a follow-up patch, when the mirror ids are set on the server-side,
the "link id"s are replaced by the mirror id of their pair.

Added llapi_layout_test tests 40-60 to exercise the layout changes.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: I65ec3246642058b5adb0791eff6940626950b076
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61962
Reviewed-by: Maximilian Dilger <mdilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>