LU-19079 sec: update nodemap membership on flavor change
When a security flavor is updated on an export (e.g. from null to skn)
the client needs to refresh or create a gss context. Now that we have
GSS based nodemaps, it is useful to lookup the nodemap entries when
this happens, to see if the client now belongs to a different nodemap,
based on its gss context.

Add sanity-sec test_105 to exercise this feature.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I8ac737a5bf91a18a499dcdd5e86e8d3fdd8fee5f
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62990
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>