LU-12668 lod: bind ec mirror to data mirror during creation
When creating a layout and allocating OSTs for the different
components, find the matching data components for parity components
and vice versa, so we know how many stripes are in the data component.
We need this in order to compute how many raidsets there will be so
that we can request to allocate one OST per parity per raidset for the
parity component. The binding is tested implicitly by existing test
that would fail if incorrect.

The additional case of inheriting the link ids between data and parity
comps when an EC template is set on a directory is also addressed
here. Because directories provide a template for EC params that new
files can inherit, directories must carry and persist unbound link ids
with their corresponding "LCME_FL_IS_LINK_ID" flag. Those are later
bound when files are created in the central binding function
"lod_bind_data_parity()". Added sanity-ec tests 1k and 1l to exercise
directory EC template inheritance.

Test-Parameters: testlist=sanity-ec
Test-Parameters: testlist=sanity-ec fstype=zfs
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I97c7586b4ee02a569f618bd6a8e124671b6dc067
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62906
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>