LU-19695 tests: fix flaky test 160s
Sanity test 160s frequently fails, especially when it follows other
changelog tests. For example, I am able to reliably get it to fail
when I run it after test 160p:

    $ ./lustre/tests/auster -s sanity --only 160p,160s

The failure seems to occur when a garbage collection thread is started
too soon. Specifically, the first metadata operation of the
test--creating the test directory--can trigger GC thread creation.

This results in an update to mc_gc_time in mdd_changelog_store():

        int mdd_changelog_store(const struct lu_env *env, struct mdd_device *mdd,
                                struct llog_changelog_rec *rec, struct thandle *th)
        {
                        ...
                        /* next check in mdd_changelog_min_gc_interval anyway */
                        mdd->mdd_cl.mc_gc_time = ktime_get_real_seconds();
                        ...
        }

Then, when the test performs metadata operations that are intended to
trigger GC, they do not, because this timestamp comparison succeeds
(when it is expected to fail):

        ...
        /* time to recover some space ?? */
        if (likely(!mdd->mdd_changelog_gc ||
                   mdd->mdd_cl.mc_gc_task != MDD_CHLG_GC_NONE ||
                   mdd->mdd_changelog_min_gc_interval >=
                        ktime_get_real_seconds() - mdd->mdd_cl.mc_gc_time))
                /* save a spin_lock trip */
                goto out_put;
        ...

This results in no GC thread being launched, when the test expects
that one would be launched.

A solution is to sleep for an additional second in the test, so that
the comparison

        mdd->mdd_changelog_min_gc_interval >= ktime_get_real_seconds() - mdd->mdd_cl.mc_gc_time

returns false, even if a GC thread was launched at the very beginning
of the test.

Test-Parameters: trivial testlist=sanity env=ONLY="160p 160s",ONLY_REPEAT=100
Signed-off-by: Thomas Bertschinger <bertschinger@lanl.gov>
Change-Id: I71037bd864fbe9dd0569692968807a8fa6861768
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63024
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>