LU-12738 mdd: enable append_stripe_count=1 by default
Re-enable mdd.*.append_stripe_count=1 by default until the client and
MDS can negotiate whether an appending write will instantiate all of
the components of the file or not.

This is needed temporarily to ensure that old clients do not cause
problems for MDS servers that do not enable append layouts by default.

Fixes: 0f21e02f17 ("LU-12738 pfl: do not instantiate full PFL layout on append")
Signed-off-by: Andreas Dilger <adilger@thelustrecollective.com>
Signed-off-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Change-Id: Ibefbb30b646671f1e91a822dc8d66b9258169811
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63067
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>