LU-19726 tests: additional output for vbr test suite failures
There are a handful of test case failures that have happened
more than once, but not frequently enough to easily replicate.
The goal of these small changes to test-framework.sh is to
add a slightly longer window for nodes to shutdown, while also
printing uptime info in case of future failures.

Signed-off-by: Maximilian Dilger <mdilger@whamcloud.com>
Change-Id: Id32add0d7762df588a51a6cb3acfc8c301dea632
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63160
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>