LU-19739 tests: quiet spurious kmemleak errors
If kmemleak is not configured, it will print error messages like:

    lustre_rmmod: line 114: echo: write error: Device or resource busy

but this can be misunderstood to indicate there is something wrong.
Quiet these error messages if kmemleak is present, but inactive, and
do not try to check for leaks every time the modules are unloaded.

Clean up code to avoid repeated full pathnames and bash code style.

Fixes: 15c0a21ea9 ("LU-12341 tests: Add kmemleak awareness to test-framework")
Signed-off-by: Andreas Dilger <adilger@thelustrecollective.com>
Change-Id: Ia108dce0474c5d09e0c46c3d90fd3d25a02540e5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63149
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>