LU-19179 build: remove RHEL7 handling of fscrypt
With the removal of RHEL7 we no longer need the LC_IS_ENCRYPTED
autoconf test so remove it. Rework the autoconf handling for
fscrypt. We handle the case of a kernel turning off fscrypt and
the case of building only the tools for the Linux native client.

Change-Id: Ic5212f297933f63ae362f2f80e632f1bb3674658
Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/63207
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>